updated data functions
